backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Database
sqlalchemy==2.0.32
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# ML Libraries
scikit-learn==1.5.1
xgboost==2.1.0
pandas==2.2.2
numpy==2.0.1
joblib==1.4.2

# Data processing
openpyxl==3.1.5
xlrd==2.0.1

# Validation
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

# Storage
boto3==1.34.158
python-dotenv==1.0.1

# Payments
stripe==10.5.0

# Background tasks
celery[redis]==5.4.0
redis==5.0.8

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0

# Utilities
aiofiles==24.1.0
